Scripts to help fix pre-requisites
I added to python scripts I made. One identifies all pre-requisites that don't exist as a course in the system and thus must be changed. The other adds courses to a students completed courses list if they've done a course that requires it as a pre-requisite. Also if they've completed a course that would be impossible to complete in the current system (pre-requisite is -1) then the course is removed from their completed courses list.